requirements.txt
flask
ollama
python-dotenv
psycopg2-binary
PyPDF2
python-pptx
diffusers
torch
transformers
accelerate